+++ src/ui/lua.tsx
@@ -1,8 +1,0 @@
-export function LuaApp() {
-  return <>
-    <div className='flex flex-col gap-3'>
-      <h2 className='h2'>Lua(init.lua)</h2>
-      <p>W.I.P.</p>
-    </div>
-  </>
-}